Passing all basic tests. Before lastKeyPressed
This version is passing all the basic tests and was uploaded before adding the major change of the lastKeyPressed functionality, which will track the last key pressed and needs to be implemented because of a bug (2 * = (res: 0, exp: 2 - or no change to main screen). 

We will need to disable the equal button if the lastKeyPressed was one of the operators.